入力フォームをディレクトリ分け
